+++ graphql/.babelrc
@@ -1,3 +1,5 @@
 {
-    "presets": [ "env"]
+     "presets": [ "env"],
+        "plugins": ["transform-object-rest-spread"]
+
 }
