+++ package.json
@@ -1,5 +1,7 @@
 {
   "dependencies": {
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "core-js": "^3.33.0",
     "graphql": "^16.8.1",
     "graphql-yoga": "^4.0.5"
   }
